requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-telegram-bot==21.6
paramiko==3.5.0
SQLAlchemy==2.0.35
python-dotenv==1.0.1
cryptography==43.0.1
pydantic==2.9.2
